+++ main.py
@@ -1,0 +1,4 @@
+from triangulation.t_gui import run_app
+
+if __name__ == "__main__":
+    run_app()
